💾 🧩 Feat, Refactor: 迁移并整理配置与设备安全相关功能的接口。
