+++ KitX Core Contracts/KitX.Core.Contract/Configuration/IActivityConfig.cs
@@ -1,0 +1,9 @@
+namespace KitX.Core.Contract.Configuration;
+
+/// <summary>
+/// Activity configuration section
+/// </summary>
+public interface IActivityConf
+{
+    int TotalRecorded { get; set; }
+}
