+++ KitX Core Contracts/KitX.Core.Contract/Configuration/IIOConfig.cs
@@ -1,0 +1,10 @@
+namespace KitX.Core.Contract.Configuration;
+
+/// <summary>
+/// IO configuration section
+/// </summary>
+public interface IIOConf
+{
+    int UpdatingCheckPerThreadFilesCount { get; set; }
+    int OperatingSystemVersionUpdateInterval { get; set; }
+}
