+++ KitX Core Contracts/KitX.Core.Contract/Configuration/ILoadersConfig.cs
@@ -1,0 +1,9 @@
+namespace KitX.Core.Contract.Configuration;
+
+/// <summary>
+/// Loaders configuration section
+/// </summary>
+public interface ILoadersConf
+{
+    string InstallPath { get; set; }
+}
