+++ KitX Core Contracts/KitX.Core.Contract/Configuration/ILogConfig.cs
@@ -1,0 +1,16 @@
+using Serilog.Events;
+
+namespace KitX.Core.Contract.Configuration;
+
+/// <summary>
+/// Log configuration section
+/// </summary>
+public interface ILogConf
+{
+    long LogFileSingleMaxSize { get; set; }
+    string LogFilePath { get; set; }
+    string LogTemplate { get; set; }
+    int LogFileMaxCount { get; set; }
+    int LogFileFlushInterval { get; set; }
+    public LogEventLevel LogLevel { get; set; }
+}
