+++ KitX Core Contracts/KitX.Core.Contract/Configuration/NavigationViewPaneDisplayMode.cs
@@ -1,0 +1,13 @@
+namespace KitX.Core.Contract.Configuration;
+
+/// <summary>
+/// Navigation view pane display mode enum
+/// </summary>
+public enum NavigationViewPaneDisplayMode
+{
+    Auto = 0,
+    Left = 1,
+    Top = 2,
+    LeftCompact = 3,
+    LeftMinimal = 4
+}
