+++ KitX Core Contracts/KitX.Core.Contract/Configuration/WindowState.cs
@@ -1,0 +1,15 @@
+using System;
+
+namespace KitX.Core.Contract.Configuration;
+
+/// <summary>
+/// Window state enumeration (mirrors Avalonia.WindowState)
+/// </summary>
+public enum WindowState
+{
+    Normal,
+    Minimized,
+    Maximized,
+    FullScreen,
+    NonInteractive
+}
